add Repertoire, create default GA algorithm, update emitter type hints and some default values
