+++ tests/core_test/containers_test/repertoire_test.py
@@ -1,7 +1,7 @@
 import jax.numpy as jnp
 import pytest
 
-from qdax.core.containers.repertoire import (
+from qdax.core.containers.mapelites_repertoire import (
     MapElitesRepertoire,
     compute_euclidean_centroids,
 )
