+++ tests/core_test/pgame_test.py
@@ -6,7 +6,10 @@ import jax.numpy as jnp
 import pytest
 
 from qdax import environments
-from qdax.core.containers.repertoire import MapElitesRepertoire, compute_cvt_centroids
+from qdax.core.containers.mapelites_repertoire import (
+    MapElitesRepertoire,
+    compute_cvt_centroids,
+)
 from qdax.core.emitters.mutation_operators import isoline_variation
 from qdax.core.emitters.pga_me_emitter import PGAMEConfig, PGEmitter
 from qdax.core.map_elites import MAPElites
